5_HandsOn_Tutorials/5.5_Docker_Containers/06_multi_service_app/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
